code/requirements.txt
absl-py==2.3.0
accelerate==1.8.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
datasets==3.6.0
defusedxml==0.7.1
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
evaluate==0.4.4
exceptiongroup==1.3.0
fairscale==0.4.13
fastapi==0.115.13
ffmpy==0.6.0
filelock==3.13.1
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
git-lfs==1.6
gradio==5.34.2
gradio_client==1.10.3
groovy==0.1.2
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
litellm==1.72.9
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
murmurhash==1.0.13
networkx==3.3
nltk==3.9.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.32.0
openai==1.90.0
opencv-python==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.3.0
peft==0.14.0
Pillow==9.5.0
platformdirs==4.3.8
preshed==3.0.10
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rouge-score==0.1.2
rpds-py==0.25.1
ruff==0.12.0
safehttpx==0.1.6
safetensors==0.5.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
supervision==0.25.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
thinc==8.3.6
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.49.0
triton==3.1.0
trl==0.17.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.158
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.34.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
